# Adopt cachecontrol 0.12.0 with msgpack support (#3515)