--- a/CHANGES.txt
+++ b/CHANGES.txt
@@ -1,11 +1,14 @@
 **9.1.0 (UNRELEASED)**
 
+* **WARNING** pip 9.1 cache format has changed. Old versions of pip cannot
+  read cache entries written by 9.1 but will handle this gracefully by
+  re-downloading the affected packages (:issue:`3515`).
+
 * Use pkg_resources to parse the entry points file to allow names with
   colons (:pull:`3901`)
 
 * Return a failing exit status when `pip install`, `pip download`, or
   `pip wheel` is called with no requirements. (:issue:`2720`, :pull:`2721`)
-
 
 * Add `--exclude-editable` to ``pip freeze`` to exclude editable packages
   from installed package list.