… pip/_vendor/msgpack/_version.py
… pip/_vendor/msgpack/fallback.py