@StephanErb 2017-01-30
I have pushed a squashed commit that replaces the `.gitignore` with a custom removal in ` tasks/vendoring/__init__.py`.