@xavfernandez 2017-01-30
Do we really get this performance boost ? I don't seem to reproduce it...
(Even if we'll still merge this as we'll want to use latest cachecontrol anyway)