@xavfernandez 2017-01-30
Hmm, my bad, I tested with both which filled the cache with a json-cached response and indeed json-cached entries are read at the same speed with the new and old cachecontrol :)